requirements.txt
pynput
SpeechRecognition[faster-whisper]
pyaudio
pystray
Pillow
pyinstaller
soundfile